Add preliminary data-gathering files
Automated the process of gathering experimental data for available CHERI
allocators. The main additions in this commit is the `get_all.py`
script, which installs, tests, and finds CHERI API calls in memory
allocators, and needed configuration to perform these operations for
each allocator in `allocators` (which should also serve as information
of which allocator versions we're testing).

Required additional `pip` modules are in `./data/README.md`. All
allocators can be run via the `allocators_folder` field in
`config.json`, or via passing a specific folder to the script via
`--alloc`.