Experiment script updates
* Added Latex table generation
* Improved support for `cheribuild` type allocators
        - ensure correct version is used
        - use explicit `.so` files for `jemalloc` and `libmalloc_simple`

All allocators should work in bulk now, except `snmalloc-cheribuild`.
There also seems to be points where the script is put in the background
and needs to be manually restarted via `fg`. It seems to happen just
before a subsequent `cheribuild` call is made. Still need to see if the
resume can be automated, or the stop prevented.