Correct analysis of FVP profile.
`malloc` and `free` are heavily affected but the others are not (in
terms of instruction count). Shared-object calls would be attributed to
"[mixed]:[other]" because .plt does not have symbols, and so won't be
obvious here. We know that calling `malloc` or `free` requires a
so-call, so binary_tree is paying a cost there, but mstress and richards
have similar performance figures but very different profiles.

Really, it looks to me like we have three very different benchmarks,
hitting different issues, and it's hard to draw conclusions from this.